refactored the code to enable adding memgraph graph backend
