+++ .env-template
@@ -1,9 +1,13 @@
 WEB_HOST=0.0.0.0
 WEB_PORT=8080
+GRAPH_DB=neo4j
 NEO4J_HOST=localhost
+NEO4J_BOLT_PORT=7687
 NEO4J_USERNAME=neo4j
-NEO4J_QUERY_TIMEOUT=600
-NEO4J_HTTP_PORT=7474
+MEMGRAPH_HOST=localhost
+MEMGRAPH_BOLT_PORT=7687
+MEMGRAPH_USERNAME=memgraph
+GRAPH_QUERY_TIMEOUT=600
 PLATER_SERVICE_ADDRESS=localhost
 PLATER_TITLE=Plater
 PLATER_VERSION=1.5.1
