+++ .gitignore
@@ -1,7 +1,7 @@
 *env
 __pycache__
 .pytest_cache
-.log
+*.log
 kubernets
 *coverage
 .idea
